+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## [0.0.12]
+* Implement `Tooltip`
 * Add mouse cursors to help button, push button and `TextField`
 
 ## [0.0.11]